refactor(byoi): extract shared k8s provider to internal/k8s-provider (P0 step 5a)
Move KubernetesProvider + buildDeploymentSpec + K8sConfig + the pure status
helpers out of control-plane into internal/k8s-provider, so the in-process
control plane and the upcoming standalone env-runner share ONE implementation
(crucially one buildDeploymentSpec -> both produce byte-identical pods). Same
relative-path + esbuild-bundle sharing pattern as internal/types.

control-plane/src/services/k8s.ts becomes a thin shim: re-exports the shared
surface and binds a process-local defaultProvider with the historical
free-function wrappers, so every `import * as k8s from '../services/k8s'` call
site is unchanged. KubernetesProvider/makeDefaultProvider are imported for cp's
own use but not re-exported — the env-runner imports them from internal directly.

tsc + golden-master snapshot (now validating the internal buildDeploymentSpec
via the shim) + 133 unit tests + esbuild bundle all green.
