+++ internal/k8s-provider/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@neutree-ai/k8s-provider",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": "./index.ts"
+  },
+  "dependencies": {
+    "@kubernetes/client-node": "0.22.3"
+  }
+}
